Move operator lint to pkg/lint package
