+++ pkg/lint/lint_test.go
@@ -14,11 +14,15 @@ import (
 	"fmt"
 	"io/fs"
 	"os"
+	"os/exec"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"sigs.k8s.io/yaml"
+
+	"github.com/redpanda-data/redpanda-operator/pkg/testutil"
 )
 
 const tagURL = "https://github.com/redpanda-data/redpanda-operator/releases/tag/"
@@ -27,6 +31,38 @@ type ChartYAML struct {
 	Version     string            `json:"version"`
 	AppVersion  string            `json:"appVersion"`
 	Annotations map[string]string `json:"annotations"`
+}
+
+func TestToolVersions(t *testing.T) {
+	golden := testutil.NewTxTar(t, "testdata/tool-versions.txtar")
+
+	for _, cmd := range []string{
+		"go version | cut -d ' ' -f 1-3",                           // cut removes os/arch
+		"goreleaser --version | awk 'NR>2 {print last} {last=$0}'", // head -n -1 doesn't work on macos
+		"helm version",
+		"k3d version",
+		"kind version | cut -d ' ' -f 1-3", // cut removes os/arch
+		"kubectl version --client=true",
+		"kustomize version",
+		"kuttl version | cut -d ' ' -f 1-7", // cut removes os/arch
+		"task --version",
+		"yq --version",
+		"helm-docs -v",
+		"ct version",
+	} {
+		out := sh(cmd)
+		bin := strings.SplitN(cmd, " ", 2)[0]
+		expect := fmt.Sprintf("# %s\n%s\n", cmd, out)
+		golden.AssertGolden(t, testutil.Text, bin, []byte(expect))
+	}
+}
+
+func sh(cmd string) string {
+	out, err := exec.Command("sh", "-c", cmd).CombinedOutput()
+	if err != nil {
+		return fmt.Sprintf("ERROR: %s\n%s", err.Error(), out)
+	}
+	return string(out)
 }
 
 func TestChartYAMLVersions(t *testing.T) {
