+++ pkg/lint/testdata/tool-versions.txtar
@@ -1,3 +1,10 @@
+-- ct --
+# ct version
+Version:	 3.9.0-4
+Git commit:	 v3.9.0-4
+Date:		 19700101-00:00:00
+License:	 Apache 2.0
+
 -- go --
 # go version | cut -d ' ' -f 1-3
 go version go1.23.2
@@ -23,6 +30,10 @@ ModuleSum:     unknown
 -- helm --
 # helm version
 version.BuildInfo{Version:"v3.16.2", GitCommit:"v3.16.2", GitTreeState:"", GoVersion:"go1.23.2"}
+
+-- helm-docs --
+# helm-docs -v
+helm-docs version 1.14.2
 
 -- k3d --
 # k3d version
